Moved files to new sub-package "models"
