+++ src/sample/Main.java
@@ -13,7 +13,7 @@ public class Main extends Application {
     @Override
     public void start(Stage primaryStage) throws Exception{
         Parent root = FXMLLoader.load(getClass().getResource("sample.fxml"));
-        primaryStage.setTitle("Opencv Testing Area");
+        primaryStage.setTitle("Set Finder");
         primaryStage.setScene(new Scene(root, 1200, 800));
         primaryStage.show();
 
@@ -22,8 +22,6 @@ public class Main extends Application {
 
     public static void main(String[] args) {
         System.loadLibrary(Core.NATIVE_LIBRARY_NAME);
-
-
         launch(args);
     }
 }
